# docs/explore/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SozoRock Place Intelligence API",
    "version": "2026-08-08",
    "description": "Versioned, non-clinical county evidence, planning-agent and collaboration contracts. Missing evidence remains missing; county evidence is never represented as ZIP-, city- or person-level evidence."
  },
  "servers": [
    {
      "url": "https://health.sozorockfoundation.org"
    }
  ],
...
  "paths": {
...
    "/api/evidence/v1/workspaces": {
      "get": {
        "tags": [
          "Workspace"
        ],
        "summary": "List authenticated workspaces.",
        "security": [
          {
            "cognitoAccessToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Membership-scoped workspaces."
          }
        }
      },
      "post": {
        "tags": [
          "Workspace"
        ],
        "summary": "Create a county workspace.",
        "security": [
          {
            "cognitoAccessToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Request rejected or evidence unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Request rejected or evidence unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Request rejected or evidence unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Request rejected or evidence unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cognitoAccessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Cognito access token"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}